Day 3 after AI review

+++ day1/puzzle1/solution.go
@@ -4,12 +4,12 @@ import (
 	"github.com/keneanung/adventofcode-2025/day1"
 )
 
-func Solve(input []string) (int, error) {
+func Solve(input []string) (int64, error) {
 	steps, err := day1.GetSteps(input)
 	if err != nil {
 		return 0, err
 	}
-	result := 0
+	result := int64(0)
 	for _, step := range steps {
 		if step.Position == 0 {
 			result++
