+++ day3/puzzle2/solution.go
@@ -1,11 +1,10 @@
 package puzzle2
 
 import (
-	"github.com/keneanung/adventofcode-2025/day3"	
+	"github.com/keneanung/adventofcode-2025/day3"
 )
 
-func Solve(input []string) (int, error) {
+func Solve(input []string) (int64, error) {
 	numberOfBatteries := 12
 	return day3.GetJoltage(input, numberOfBatteries)
 }
-
